+++ pkg/controller/user/importbatch.go
@@ -15,9 +15,7 @@
 package user
 
 import (
-	"context"
 	"errors"
-	"fmt"
 	"net/http"
 
 	"github.com/google/exposure-notifications-verification-server/pkg/api"
@@ -106,49 +104,4 @@ func (c *Controller) HandleImportBatch() http.Handler {
 
 		c.h.RenderJSON(w, http.StatusOK, response)
 	})
-}
-
-func (c *Controller) sendInvitation(ctx context.Context, toEmail string) error {
-
-	if err := c.sendInvitationFromRealmEmailer(ctx, toEmail); err == nil {
-		return nil
-	}
-
-	// Fallback to Firebase
-
-	if err := c.firebaseInternal.SendNewUserInvitation(ctx, toEmail); err != nil {
-		c.logger.Warnw("failed sending invitation", "error", err)
-		return fmt.Errorf("failed sending invitation: %w", err)
-	}
-	return nil
-}
-
-func (c *Controller) sendInvitationFromRealmEmailer(ctx context.Context, toEmail string) error {
-	// Send email with realm email config
-	realm := controller.RealmFromContext(ctx)
-	if realm == nil {
-		return errors.New("no realm found")
-	}
-
-	emailer, err := realm.EmailProvider(c.db)
-	if emailer == nil {
-		return errors.New("no emailer found")
-	}
-	if err != nil {
-		c.logger.Warnw("failed to get emailer for realm:", "error", err)
-		return fmt.Errorf("failed to get emailer for realm: %w", err)
-	}
-
-	message, err := controller.ComposeInviteEmail(ctx, c.h, c.client, toEmail, emailer.From(), realm.Name)
-	if err != nil {
-		c.logger.Warnw("failed composing invitation", "error", err)
-		return fmt.Errorf("failed composing invitation: %w", err)
-	}
-
-	if err := emailer.SendEmail(ctx, toEmail, message); err != nil {
-		c.logger.Warnw("failed sending invitation", "error", err)
-		return fmt.Errorf("failed sending invitation: %w", err)
-	}
-
-	return nil
 }
